:recycle: [Refactor]: sfu gateway 통합
