+++ apps/media/src/sfu/interfaces/room-transport-info.interface.ts
@@ -1,0 +1,6 @@
+import * as mediasoup from 'mediasoup';
+
+export interface RoomTransportInfo {
+  producerTransportId?: string;
+  transports: Map<string, mediasoup.types.WebRtcTransport>;
+}
