+++ apps/media/src/sfu/interfaces/transport-params.interface.ts
@@ -1,0 +1,7 @@
+import * as mediasoup from 'mediasoup';
+
+export interface TransportParams {
+  roomId: string;
+  transportId: string;
+  dtlsParameters: mediasoup.types.DtlsParameters;
+}
